refactor(app): add hook for overflow menu outside click
This PR abstracts the logic to hide the overflow menu when an outside node is clicked.

closes #10086
