+++ app/src/organisms/Devices/RobotOverflowMenu.tsx
@@ -10,7 +10,6 @@ import {
   DIRECTION_COLUMN,
   POSITION_RELATIVE,
   ALIGN_FLEX_END,
-  Overlay,
   SPACING,
   TEXT_TRANSFORM_CAPITALIZE,
 } from '@opentrons/components'
@@ -18,10 +17,10 @@ import { CONNECTABLE, removeRobot } from '../../redux/discovery'
 import { OverflowBtn } from '../../atoms/MenuList/OverflowBtn'
 import { Divider } from '../../atoms/structure'
 import { MenuItem } from '../../atoms/MenuList/MenuItem'
-import { Portal } from '../../App/portal'
 import { ChooseProtocolSlideout } from '../ChooseProtocolSlideout'
 import { useCurrentRunId } from '../ProtocolUpload/hooks'
 import { ConnectionTroubleshootingModal } from './ConnectionTroubleshootingModal'
+import { useMenuHandleClickOutside } from '../../atoms/MenuList/hooks'
 
 import type { StyleProps } from '@opentrons/components'
 import type { DiscoveredRobot } from '../../redux/discovery/types'
@@ -33,7 +32,12 @@ interface RobotOverflowMenuProps extends StyleProps {
 export function RobotOverflowMenu(props: RobotOverflowMenuProps): JSX.Element {
   const { robot, ...styleProps } = props
   const { t } = useTranslation(['devices_landing', 'shared'])
-  const [showOverflowMenu, setShowOverflowMenu] = React.useState<boolean>(false)
+  const {
+    MenuOverlayPortal,
+    handleOverflowClick,
+    showOverflowMenu,
+    setShowOverflowMenu,
+  } = useMenuHandleClickOutside()
   const dispatch = useDispatch<Dispatch>()
   const runId = useCurrentRunId()
   const [
@@ -49,14 +53,6 @@ export function RobotOverflowMenu(props: RobotOverflowMenuProps): JSX.Element {
     e.preventDefault()
     setShowChooseProtocolSlideout(true)
     setShowOverflowMenu(!showOverflowMenu)
-  }
-  const handleOverflowClick: React.MouseEventHandler<HTMLButtonElement> = e => {
-    e.preventDefault()
-    setShowOverflowMenu(!showOverflowMenu)
-  }
-  const handleClickOutside: React.MouseEventHandler<HTMLDivElement> = e => {
-    e.preventDefault()
-    setShowOverflowMenu(false)
   }
   const handleClickConnectionTroubleshooting: React.MouseEventHandler<HTMLButtonElement> = e => {
     e.preventDefault()
@@ -144,26 +140,19 @@ export function RobotOverflowMenu(props: RobotOverflowMenuProps): JSX.Element {
           {menuItems}
         </Flex>
       ) : null}
-      <Portal level="top">
-        {showOverflowMenu ? (
-          <Overlay
-            onClick={handleClickOutside}
-            backgroundColor={COLORS.transparent}
-          />
-        ) : null}
-        {robot.status === CONNECTABLE ? (
-          <ChooseProtocolSlideout
-            robot={robot}
-            showSlideout={showChooseProtocolSlideout}
-            onCloseClick={() => setShowChooseProtocolSlideout(false)}
-          />
-        ) : null}
-        {showConnectionTroubleshootingModal ? (
-          <ConnectionTroubleshootingModal
-            onClose={() => setShowConnectionTroubleshootingModal(false)}
-          />
-        ) : null}
-      </Portal>
+      <MenuOverlayPortal />
+      {robot.status === CONNECTABLE ? (
+        <ChooseProtocolSlideout
+          robot={robot}
+          showSlideout={showChooseProtocolSlideout}
+          onCloseClick={() => setShowChooseProtocolSlideout(false)}
+        />
+      ) : null}
+      {showConnectionTroubleshootingModal ? (
+        <ConnectionTroubleshootingModal
+          onClose={() => setShowConnectionTroubleshootingModal(false)}
+        />
+      ) : null}
     </Flex>
   )
 }
