+++ app/src/organisms/ProtocolDetails/ProtocolLabwareDetails.tsx
@@ -8,7 +8,6 @@ import {
   Flex,
   Icon,
   JUSTIFY_SPACE_BETWEEN,
-  Overlay,
   POSITION_ABSOLUTE,
   POSITION_RELATIVE,
   SIZE_5,
@@ -20,8 +19,8 @@ import { StyledText } from '../../atoms/text'
 import { Divider } from '../../atoms/structure'
 import { OverflowBtn } from '../../atoms/MenuList/OverflowBtn'
 import { MenuItem } from '../../atoms/MenuList/MenuItem'
-import { Portal } from '../../App/portal'
 import { LabwareDetails } from '../Labware/LabwareDetails'
+import { useMenuHandleClickOutside } from '../../atoms/MenuList/hooks'
 
 import type { LoadLabwareRunTimeCommand } from '@opentrons/shared-data/protocol/types/schemaV6/command/setup'
 import type { LabwareDefAndDate } from '../Labware/hooks'
@@ -144,21 +143,16 @@ export const LabwareDetailOverflowMenu = (
 ): JSX.Element => {
   const { labware } = props
   const { t } = useTranslation('protocol_details')
-  const [showOverflowMenu, setShowOverflowMenu] = React.useState<boolean>(false)
+  const {
+    MenuOverlayPortal,
+    handleOverflowClick,
+    showOverflowMenu,
+    setShowOverflowMenu,
+  } = useMenuHandleClickOutside()
   const [
     showLabwareDetailSlideout,
     setShowLabwareDetailSlideout,
   ] = React.useState<boolean>(false)
-
-  const handleOverflowClick: React.MouseEventHandler<HTMLButtonElement> = e => {
-    e.preventDefault()
-    setShowOverflowMenu(!showOverflowMenu)
-  }
-
-  const handleClickOutside: React.MouseEventHandler<HTMLDivElement> = e => {
-    e.preventDefault()
-    setShowOverflowMenu(false)
-  }
 
   const handleClickMenuItem: React.MouseEventHandler<HTMLButtonElement> = e => {
     e.preventDefault()
@@ -192,20 +186,13 @@ export const LabwareDetailOverflowMenu = (
           </MenuItem>
         </Flex>
       ) : null}
-      <Portal level="top">
-        {showOverflowMenu ? (
-          <Overlay
-            onClick={handleClickOutside}
-            backgroundColor={COLORS.transparent}
-          />
-        ) : null}
-        {showLabwareDetailSlideout ? (
-          <LabwareDetails
-            labware={labware}
-            onClose={() => setShowLabwareDetailSlideout(false)}
-          />
-        ) : null}
-      </Portal>
+      <MenuOverlayPortal />
+      {showLabwareDetailSlideout ? (
+        <LabwareDetails
+          labware={labware}
+          onClose={() => setShowLabwareDetailSlideout(false)}
+        />
+      ) : null}
     </Flex>
   )
 }
